# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FDC+ Serial Drive Server API",
    "version": "2.0.0",
    "description": "REST API for the Altair 8800 FDC+ Serial Drive Server web interface. Manages disk image serving, CP/M filesystem browsing, cassette audio, terminal serial connections, script replay, and XMODEM transfers.\n\n## Socket.IO Events\n\nThe server also exposes real-time events via Socket.IO (not covered by OpenAPI).\n\n### Server → Client\n\n| Event | Payload | Description |\n|-------|---------|-------------|\n| `status` | `ServerStatus` | Periodic server status broadcast (1s interval) |\n| `terminal:data` | `number[]` | Incoming serial data as byte array |\n| `terminal:status` | `TerminalStatus` | Terminal connection status change |\n| `terminal:error` | `{message: string}` | Terminal error |\n| `replay:progress` | `ReplayProgress` | Transfer progress update |\n| `replay:status` | `{active, mode, progress}` | Transfer status on connect |\n\n### Client → Server\n\n| Event | Payload | Description |\n|-------|---------|-------------|\n| `request-status` | _(none)_ | Request current server status |\n| `terminal:write` | `string` | Send keyboard input to terminal serial port |\n| `terminal:control` | `{type, value}` | Set DTR/RTS control signals |\n| `replay:start` | `ReplayStartParams` | Start raw replay or XMODEM send |\n| `replay:cancel` | _(none)_ | Cancel active transfer |"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
...
    }
  ],
...
  "paths": {
...
    "/api/clients/{clientId}/name": {
      "put": {
        "tags": [
          "Clients"
        ],
        "summary": "Set a client's friendly name",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Name updated"
          }
        }
      }
    },
    "/api/clients/{clientId}/drives/{drive}": {
      "put": {
        "tags": [
          "Clients"
        ],
        "summary": "Set a client's per-drive mount override",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "drive",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "filename"
                ],
                "properties": {
                  "filename": {
                    "type": "string"
                  },
                  "readonly": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Override set"
          },
          "404": {
            "description": "Image not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Clients"
        ],
        "summary": "Clear a client's per-drive override (inherit global)",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "drive",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Override cleared"
          }
        }
      }
    },
    "/api/clients/{clientId}": {
      "delete": {
        "tags": [
          "Clients"
        ],
        "summary": "Forget a client (clear its overrides, splinters, and name)",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client forgotten"
          }
        }
      }
    },
    "/api/clients/{clientId}/drives/{drive}/splinter/commit": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Commit a client's splinter back onto its master image (hot-swap)",
        "description": "Writes the client's private copy-on-write splinter for this drive back onto the shared master image, then hot-reloads every open handle (operator drives via a swap window, client sessions via resync) so live readers pick up the new contents. Client splinters re-attach by base name, keeping their own writes. Refused only when the base is held read-write by a live master-write path (an operator drive mounted read-write, or the connected master-write client), whose in-flight write would be clobbered.",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "drive",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Splinter committed to the master; response includes hotSwapped and the reloadedDrives list"
          },
          "404": {
            "description": "No splinter for this client/drive",
            "or master missing": null
          },
          "409": {
            "description": "Base is held read-write by a live master-write path"
          }
        }
      }
    },
    "/api/clients/{clientId}/drives/{drive}/splinter/save-snapshot": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Save a client's splinter as a snapshot of its master",
        "description": "Captures the client's current copy-on-write splinter for this drive as a snapshot of the master image, without touching the master or the splinter.",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "drive",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Snapshot saved"
          },
          "404": {
            "description": "No splinter for this client/drive"
          }
        }
      }
    },
    "/api/clients/{clientId}/drives/{drive}/splinter/save-as-disk": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Save a client's splinter as a new named disk image",
        "description": "Non-destructive publish — copies the client's copy-on-write splinter for this drive to a new named image in the disk library, without touching the live master. The name is suffixed (-2, -3, …) on collision; the extension defaults to the master's if omitted.",
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "drive",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved; response includes the new filename"
          },
          "400": {
            "description": "Invalid filename or extension"
          },
          "404": {
            "description": "No splinter for this client/drive"
          }
        }
      }
    },
...
  }
}